+++ selfdrive/camerad/cameras/camera_common.h
@@ -43,8 +43,9 @@ const bool env_send_road = getenv("SEND_ROAD") != NULL;
 const bool env_send_wide_road = getenv("SEND_WIDE_ROAD") != NULL;
 
 // for debugging
-// note: ONLY_ROAD doesn't work, likely due to a mixup with wideRoad cam in the kernel
-const bool env_only_driver = getenv("ONLY_DRIVER") != NULL;
+const bool env_disable_road = getenv("DISABLE_ROAD") != NULL;
+const bool env_disable_wide_road = getenv("DISABLE_WIDE_ROAD") != NULL;
+const bool env_disable_driver = getenv("DISABLE_DRIVER") != NULL;
 const bool env_debug_frames = getenv("DEBUG_FRAMES") != NULL;
 
 typedef void (*release_cb)(void *cookie, int buf_idx);