+++ selfdrive/car/hyundai/carcontroller.py
@@ -88,17 +88,18 @@ class CarController:
           self.last_resume_frame = self.frame
 
     if self.frame % 2 == 0 and self.CP.openpilotLongitudinalControl:
-      lead_visible = False
-      accel = actuators.accel if CC.longActive else 0
+      accel = actuators.accel
+      jerk = 0
 
-      jerk = clip(2.0 * (accel - CS.out.aEgo), -12.7, 12.7)
-
-      if accel < 0:
-        accel = interp(accel - CS.out.aEgo, [-1.0, -0.5], [2 * accel, accel])
+      if CC.longActive:
+        jerk = clip(2.0 * (accel - CS.out.aEgo), -12.7, 12.7)
+        if accel < 0:
+          accel = interp(accel - CS.out.aEgo, [-1.0, -0.5], [2 * accel, accel])
 
       accel = clip(accel, CarControllerParams.ACCEL_MIN, CarControllerParams.ACCEL_MAX)
 
-      stopping = (actuators.longControlState == LongCtrlState.stopping)
+      lead_visible = False
+      stopping = actuators.longControlState == LongCtrlState.stopping
       set_speed_in_units = hud_control.setSpeed * (CV.MS_TO_MPH if CS.clu11["CF_Clu_SPEED_UNIT"] == 1 else CV.MS_TO_KPH)
       can_sends.extend(create_acc_commands(self.packer, CC.enabled, accel, jerk, int(self.frame / 2), lead_visible,
                                            set_speed_in_units, stopping, CS.out.gasPressed))