+++ selfdrive/hardware/base.py
@@ -134,9 +134,11 @@ class HardwareBase(ABC):
   def get_gpu_usage_percent(self):
     pass
 
-  @abstractmethod
   def get_modem_version(self):
-    pass
+    return None
+
+  def get_modem_nv(self):
+    return None
 
   @abstractmethod
   def get_modem_temperatures(self):
@@ -148,6 +150,9 @@ class HardwareBase(ABC):
 
   @abstractmethod
   def initialize_hardware(self):
+    pass
+
+  def configure_modem(self):
     pass
 
   @abstractmethod