+++ selfdrive/test/process_replay/process_replay.py
@@ -346,6 +346,7 @@ def setup_env(simulation=False):
   params.clear_all()
   params.put_bool("OpenpilotEnabledToggle", True)
   params.put_bool("Passive", False)
+  params.put_bool("DisengageOnAccelerator", True)
 
   os.environ["NO_RADAR_SLEEP"] = "1"
   os.environ["REPLAY"] = "1"
@@ -492,3 +493,14 @@ def cpp_replay_process(cfg, lr, fingerprint=None):
     managed_processes[cfg.proc_name].stop()
 
   return log_msgs
+
+
+def check_enabled(msgs):
+  for msg in msgs:
+    if msg.which() == "carParams":
+      if msg.carParams.notCar:
+        return True
+    elif msg.which() == "controlsState":
+      if msg.controlsState.active:
+        return True
+  return False